All pages have nav menu
I updated every pages to have the nav menu and can then go back to the home page if interested. I did not put the nav menu on the parking lot screen, because there is already a lot happening on that page.  I deleted some files that were not used as well.